Move DtStatus and DtStraightPathFlags to lib.rs and use recastnavigation-sys as implementation.
